Make the route controller as a separate binary
