+++ controllers/main.go
@@ -349,6 +349,7 @@ func main() {
 				os.Exit(1)
 			}
 		}
+
 	}
 
 	// Setup webhooks with manager
